Move some WorkspaceFileResolver functions into LSPPlatform
