+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Deprecated `luau-lsp.types.roblox` setting in favour of `luau-lsp.platform.type`
 - Deprecated `luau-lsp.completion.imports.suggestServices` setting in favour of `luau-lsp.platform.roblox.suggestServices`
+- Deprecated `luau-lsp.sourcemap.*` settings in favour of `luau-lsp.platform.roblox.sourcemap.*`
 
 ### Added
