+++ editors/code/package.json
@@ -158,27 +158,67 @@
           "markdownDescription": "Whether Rojo sourcemap parsing is enabled",
           "type": "boolean",
           "default": true,
-          "scope": "resource"
+          "scope": "resource",
+          "markdownDeprecationMessage": "**Deprecated**: Please use `#luau-lsp.platform.roblox.sourcemap.enabled#` instead.",
+          "deprecationMessage": "Deprecated: Please use luau-lsp.platform.roblox.sourcemap.enabled instead."
         },
         "luau-lsp.sourcemap.autogenerate": {
           "markdownDescription": "Automatically run the `rojo sourcemap` command to regenerate sourcemaps on changes",
           "type": "boolean",
           "default": true,
-          "scope": "resource"
+          "scope": "resource",
+          "markdownDeprecationMessage": "**Deprecated**: Please use `#luau-lsp.platform.roblox.sourcemap.autogenerate#` instead.",
+          "deprecationMessage": "Deprecated: Please use luau-lsp.platform.roblox.sourcemap.autogenerate instead."
         },
         "luau-lsp.sourcemap.rojoPath": {
           "markdownDescription": "Path to the Rojo executable. If not provided, attempts to run `rojo` in the workspace directory, so it must be available on the PATH",
           "type": "string",
           "default": null,
-          "scope": "resource"
+          "scope": "resource",
+          "markdownDeprecationMessage": "**Deprecated**: Please use `#luau-lsp.platform.roblox.sourcemap.rojoPath#` instead.",
+          "deprecationMessage": "Deprecated: Please use luau-lsp.platform.roblox.sourcemap.rojoPath instead."
         },
         "luau-lsp.sourcemap.rojoProjectFile": {
           "markdownDescription": "The name of the Rojo project file to generate a sourcemap for.\nOnly applies if `#luau-lsp.sourcemap.autogenerate#` is enabled",
           "type": "string",
           "default": "default.project.json",
-          "scope": "resource"
+          "scope": "resource",
+          "markdownDeprecationMessage": "**Deprecated**: Please use `#luau-lsp.platform.roblox.sourcemap.rojoProjectFile#` instead.",
+          "deprecationMessage": "Deprecated: Please use luau-lsp.platform.roblox.sourcemap.rojoProjectFile instead."
         },
         "luau-lsp.sourcemap.includeNonScripts": {
+          "markdownDescription": "Include non-script instances in the generated sourcemap",
+          "type": "boolean",
+          "default": true,
+          "scope": "resource",
+          "markdownDeprecationMessage": "**Deprecated**: Please use `#luau-lsp.platform.roblox.sourcemap.includeNonScripts#` instead.",
+          "deprecationMessage": "Deprecated: Please use luau-lsp.platform.roblox.sourcemap.includeNonScripts instead."
+        },
+        "luau-lsp.platform.roblox.sourcemap.enabled": {
+          "markdownDescription": "Whether Rojo sourcemap parsing is enabled",
+          "type": "boolean",
+          "default": true,
+          "scope": "resource"
+        },
+        "luau-lsp.platform.roblox.sourcemap.autogenerate": {
+          "markdownDescription": "Automatically run the `rojo sourcemap` command to regenerate sourcemaps on changes",
+          "type": "boolean",
+          "default": true,
+          "scope": "resource"
+        },
+        "luau-lsp.platform.roblox.sourcemap.rojoPath": {
+          "markdownDescription": "Path to the Rojo executable. If not provided, attempts to run `rojo` in the workspace directory, so it must be available on the PATH",
+          "type": "string",
+          "default": null,
+          "scope": "resource"
+        },
+        "luau-lsp.platform.roblox.sourcemap.rojoProjectFile": {
+          "markdownDescription": "The name of the Rojo project file to generate a sourcemap for.\nOnly applies if `#luau-lsp.sourcemap.autogenerate#` is enabled",
+          "type": "string",
+          "default": "default.project.json",
+          "scope": "resource"
+        },
+        "luau-lsp.platform.roblox.sourcemap.includeNonScripts": {
           "markdownDescription": "Include non-script instances in the generated sourcemap",
           "type": "boolean",
           "default": true,
@@ -221,6 +261,14 @@
           "scope": "resource"
         },
         "luau-lsp.diagnostics.strictDatamodelTypes": {
+          "markdownDescription": "Use strict DataModel types in diagnostics. When on, this is equivalent to the more expressive autocompletion types. When this is off, `game`/`script`/`workspace` (and their members) are all typed as `any`, and helps to prevent false positives. [Read More](https://github.com/JohnnyMorganz/luau-lsp/issues/83#issuecomment-1192865024)",
+          "type": "boolean",
+          "default": false,
+          "scope": "resource",
+          "markdownDeprecationMessage": "**Deprecated**: Please use `#luau-lsp.platform.roblox.diagnostics.strictDatamodelTypes#` instead.",
+          "deprecationMessage": "Deprecated: Please use luau-lsp.platform.roblox.diagnostics.strictDatamodelTypes instead."
+        },
+        "luau-lsp.platform.roblox.diagnostics.strictDatamodelTypes": {
           "markdownDescription": "Use strict DataModel types in diagnostics. When on, this is equivalent to the more expressive autocompletion types. When this is off, `game`/`script`/`workspace` (and their members) are all typed as `any`, and helps to prevent false positives. [Read More](https://github.com/JohnnyMorganz/luau-lsp/issues/83#issuecomment-1192865024)",
           "type": "boolean",
           "default": false,
