+++ src/include/LSP/WorkspaceFileResolver.hpp
@@ -8,8 +8,8 @@
 #include "Luau/Config.h"
 #include "LSP/Client.hpp"
 #include "LSP/Uri.hpp"
-#include "LSP/Sourcemap.hpp"
 #include "LSP/TextDocument.hpp"
+#include "Platform/LSPPlatform.hpp"
 
 
 // A wrapper around a text document pointer
@@ -68,25 +68,19 @@ public:
     }
 };
 
-std::optional<std::filesystem::path> resolveDirectoryAlias(
-    const std::filesystem::path& rootPath, const std::unordered_map<std::string, std::string>& directoryAliases, const std::string& str);
-
 struct WorkspaceFileResolver
     : Luau::FileResolver
     , Luau::ConfigResolver
 {
 private:
-    mutable std::unordered_map<std::string, SourceNodePtr> realPathsToSourceNodes{};
     mutable std::unordered_map<std::string, Luau::Config> configCache{};
 
 public:
     Luau::Config defaultConfig;
     std::shared_ptr<BaseClient> client;
     Uri rootUri;
-    // The root source node from a parsed Rojo source map
-    SourceNodePtr rootSourceNode;
 
-    mutable std::unordered_map<Luau::ModuleName, SourceNodePtr> virtualPathsToSourceNodes{};
+    LSPPlatform* platform = nullptr;
 
     // Currently opened files where content is managed by client
     mutable std::unordered_map</* DocumentUri */ std::string, TextDocument> managedFiles{};
@@ -109,38 +103,16 @@ public:
 
     TextDocumentPtr getOrCreateTextDocumentFromModuleName(const Luau::ModuleName& name);
 
-    /// The name points to a virtual path (i.e., game/ or ProjectRoot/)
-    static bool isVirtualPath(const Luau::ModuleName& name)
-    {
-        return name == "game" || name == "ProjectRoot" || Luau::startsWith(name, "game/") || Luau::startsWith(name, "ProjectRoot/");
-    }
-
-    std::filesystem::path getRequireBasePath(std::optional<Luau::ModuleName> fileModuleName) const;
-
-    std::optional<Luau::ModuleInfo> resolveStringRequire(const Luau::ModuleInfo* context, const std::string& requiredString) const;
-
     // Return the corresponding module name from a file Uri
     // We first try and find a virtual file path which matches it, and return that. Otherwise, we use the file system path
     Luau::ModuleName getModuleName(const Uri& name) const;
-
-    std::optional<SourceNodePtr> getSourceNodeFromVirtualPath(const Luau::ModuleName& name) const;
-
-    std::optional<SourceNodePtr> getSourceNodeFromRealPath(const std::string& name) const;
-
-    std::optional<std::filesystem::path> getRealPathFromSourceNode(const SourceNodePtr& sourceNode) const;
-
-    std::optional<std::filesystem::path> resolveToRealPath(const Luau::ModuleName& name) const;
 
     std::optional<Luau::SourceCode> readSource(const Luau::ModuleName& name) override;
     std::optional<Luau::ModuleInfo> resolveModule(const Luau::ModuleInfo* context, Luau::AstExpr* node) override;
     std::string getHumanReadableModuleName(const Luau::ModuleName& name) const override;
     const Luau::Config& getConfig(const Luau::ModuleName& name) const override;
     void clearConfigCache();
-    void updateSourceMap(const std::string& sourceMapContents);
 
 private:
-    static Luau::ModuleName getVirtualPathFromSourceNode(const SourceNodePtr& sourceNode);
-    std::optional<Luau::ModuleName> resolveToVirtualPath(const std::string& name) const;
     const Luau::Config& readConfigRec(const std::filesystem::path& path) const;
-    void writePathsToMap(const SourceNodePtr& node, const std::string& base);
 };
