+++ src/include/Platform/LSPPlatform.hpp
@@ -1,42 +1,66 @@
 #pragma once
 
 #include "LSP/ClientConfiguration.hpp"
-#include "LSP/LuauExt.hpp"
 #include "LSP/TextDocument.hpp"
-#include "LSP/WorkspaceFileResolver.hpp"
 #include "Luau/Ast.h"
 #include "Luau/Autocomplete.h"
+#include "Luau/FileResolver.h"
 #include "Luau/Frontend.h"
 #include "Luau/GlobalTypes.h"
 #include "Luau/Module.h"
 #include "Luau/TypeFwd.h"
+#include "Protocol/CodeAction.hpp"
 #include "Protocol/Completion.hpp"
 #include "Protocol/LanguageFeatures.hpp"
 #include "Protocol/SignatureHelp.hpp"
+#include "Protocol/Workspace.hpp"
 #include "nlohmann/json.hpp"
 
+#include <filesystem>
 #include <memory>
 #include <unordered_set>
 
 class WorkspaceFolder;
+struct WorkspaceFileResolver;
 
 class LSPPlatform
 {
 protected:
+    WorkspaceFileResolver* fileResolver;
     WorkspaceFolder* workspaceFolder;
 
 public:
-    virtual std::unique_ptr<FindImportsVisitor> getImportVisitor()
-    {
-        return std::make_unique<FindImportsVisitor>();
-    }
-
     virtual void mutateRegisteredDefinitions(Luau::GlobalTypes& globals, std::optional<nlohmann::json> metadata) {}
 
-    virtual void handleSourcemapUpdate(
-        Luau::Frontend& frontend, const Luau::GlobalTypes& globals, const WorkspaceFileResolver& fileResolver, bool expressiveTypes)
+    virtual void onDidChangeWatchedFiles(const lsp::FileEvent& change) {}
+
+    virtual void setupWithConfiguration(const ClientConfiguration& config) {}
+
+    /// The name points to a virtual path (i.e. for Roblox, game/ or ProjectRoot/)
+    [[nodiscard]] virtual bool isVirtualPath(const Luau::ModuleName& name) const
     {
+        return false;
     }
+
+    [[nodiscard]] virtual std::optional<Luau::ModuleName> resolveToVirtualPath(const std::string& name) const
+    {
+        return std::nullopt;
+    }
+
+    [[nodiscard]] virtual std::optional<std::filesystem::path> resolveToRealPath(const Luau::ModuleName& name) const
+    {
+        return name;
+    }
+
+    [[nodiscard]] virtual Luau::SourceCode::Type sourceCodeTypeFromPath(const std::filesystem::path& path) const
+    {
+        return Luau::SourceCode::Type::Module;
+    }
+
+    [[nodiscard]] virtual std::optional<std::string> readSourceCode(const Luau::ModuleName& name, const std::filesystem::path& path) const;
+
+    std::optional<Luau::ModuleInfo> resolveStringRequire(const Luau::ModuleInfo* context, const std::string& requiredString);
+    virtual std::optional<Luau::ModuleInfo> resolveModule(const Luau::ModuleInfo* context, Luau::AstExpr* node);
 
     virtual void handleCompletion(
         const TextDocument& textDocument, const Luau::SourceModule& module, Luau::Position position, std::vector<lsp::CompletionItem>& items)
@@ -44,10 +68,7 @@ public:
     }
 
     virtual std::optional<Luau::AutocompleteEntryMap> completionCallback(
-        const std::string& tag, std::optional<const Luau::ClassType*> ctx, std::optional<std::string> contents, const Luau::ModuleName& moduleName)
-    {
-        return std::nullopt;
-    }
+        const std::string& tag, std::optional<const Luau::ClassType*> ctx, std::optional<std::string> contents, const Luau::ModuleName& moduleName);
 
     virtual const char* handleSortText(
         const Luau::Frontend& frontend, const std::string& name, const Luau::AutocompleteEntry& entry, const std::unordered_set<std::string>& tags)
@@ -60,13 +81,8 @@ public:
         return std::nullopt;
     }
 
-    virtual void handleSuggestImports(const ClientConfiguration& config, FindImportsVisitor* importsVisitor, size_t hotCommentsLineNumber,
-        bool completingTypeReferencePrefix, std::vector<lsp::CompletionItem>& items)
-    {
-    }
-
-    virtual void handleRequireAutoImport(const std::string& requirePath, size_t lineNumber, bool isRelative, const ClientConfiguration& config,
-        FindImportsVisitor* importsVisitor, size_t hotCommentsLineNumber, std::vector<lsp::TextEdit>& textEdits)
+    virtual void handleSuggestImports(const TextDocument& textDocument, const Luau::SourceModule& module, const ClientConfiguration& config,
+        size_t hotCommentsLineNumber, bool completingTypeReferencePrefix, std::vector<lsp::CompletionItem>& items)
     {
     }
 
@@ -97,13 +113,20 @@ public:
         return false;
     }
 
-    static std::unique_ptr<LSPPlatform> getPlatform(const ClientConfiguration& config, WorkspaceFolder* workspaceFolder = nullptr);
+    static std::unique_ptr<LSPPlatform> getPlatform(
+        const ClientConfiguration& config, WorkspaceFileResolver* fileResolver, WorkspaceFolder* workspaceFolder = nullptr);
 
     LSPPlatform(const LSPPlatform& copyFrom) = delete;
     LSPPlatform(LSPPlatform&&) = delete;
     LSPPlatform& operator=(const LSPPlatform& copyFrom) = delete;
     LSPPlatform& operator=(LSPPlatform&&) = delete;
 
-    LSPPlatform(WorkspaceFolder* workspaceFolder = nullptr);
+    LSPPlatform(WorkspaceFileResolver* fileResolver = nullptr, WorkspaceFolder* workspaceFolder = nullptr);
     virtual ~LSPPlatform() = default;
+
+private:
+    [[nodiscard]] std::filesystem::path getRequireBasePath(std::optional<Luau::ModuleName> fileModuleName) const;
 };
+
+std::optional<std::filesystem::path> resolveDirectoryAlias(
+    const std::filesystem::path& rootPath, const std::unordered_map<std::string, std::string>& directoryAliases, const std::string& str);
