+++ src/operations/Diagnostics.cpp
@@ -45,7 +45,7 @@ lsp::DocumentDiagnosticReport WorkspaceFolder::documentDiagnostics(const lsp::Do
         }
         else
         {
-            auto fileName = fileResolver.resolveToRealPath(error.moduleName);
+            auto fileName = platform->resolveToRealPath(error.moduleName);
             if (!fileName || isIgnoredFile(*fileName, config))
                 continue;
             auto textDocument = fileResolver.getTextDocumentFromModuleName(error.moduleName);
@@ -83,6 +83,12 @@ lsp::DocumentDiagnosticReport WorkspaceFolder::documentDiagnostics(const lsp::Do
 
 lsp::WorkspaceDiagnosticReport WorkspaceFolder::workspaceDiagnostics(const lsp::WorkspaceDiagnosticParams& params)
 {
+    if (!isConfigured)
+    {
+        lsp::DiagnosticServerCancellationData cancellationData{/*retriggerRequest: */ true};
+        throw JsonRpcException(lsp::ErrorCode::ServerCancelled, "server not yet received configuration for diagnostics", cancellationData);
+    }
+
     lsp::WorkspaceDiagnosticReport workspaceReport;
 
     // Don't compute any workspace diagnostics for null workspace
@@ -161,6 +167,56 @@ lsp::DocumentDiagnosticReport LanguageServer::documentDiagnostic(const lsp::Docu
 {
     auto workspace = findWorkspace(params.textDocument.uri);
     return workspace->documentDiagnostics(params);
+}
+
+void WorkspaceFolder::pushDiagnostics(const lsp::DocumentUri& uri, const size_t version)
+{
+    // Convert the diagnostics report into a series of diagnostics published for each relevant file
+    lsp::DocumentDiagnosticParams params{lsp::TextDocumentIdentifier{uri}};
+    auto diagnostics = documentDiagnostics(params);
+    client->publishDiagnostics(lsp::PublishDiagnosticsParams{uri, version, diagnostics.items});
+
+    if (!diagnostics.relatedDocuments.empty())
+    {
+        for (const auto& [relatedUri, relatedDiagnostics] : diagnostics.relatedDocuments)
+        {
+            if (relatedDiagnostics.kind == lsp::DocumentDiagnosticReportKind::Full)
+            {
+                client->publishDiagnostics(lsp::PublishDiagnosticsParams{Uri::parse(relatedUri), std::nullopt, relatedDiagnostics.items});
+            }
+        }
+    }
+}
+
+/// Recompute all necessary diagnostics when we detect a configuration (or sourcemap) change
+void WorkspaceFolder::recomputeDiagnostics(const ClientConfiguration& config) {
+    // Handle diagnostics if in push-mode
+    if ((!client->capabilities.textDocument || !client->capabilities.textDocument->diagnostic))
+    {
+        // Recompute workspace diagnostics if requested
+        if (config.diagnostics.workspace)
+        {
+            auto diagnostics = workspaceDiagnostics({});
+            for (const auto& report : diagnostics.items)
+            {
+                if (report.kind == lsp::DocumentDiagnosticReportKind::Full)
+                {
+                    client->publishDiagnostics(lsp::PublishDiagnosticsParams{report.uri, report.version, report.items});
+                }
+            }
+        }
+        // Recompute diagnostics for all currently opened files
+        else
+        {
+            for (const auto& [_, document] : fileResolver.managedFiles)
+                pushDiagnostics(document.uri(), document.version());
+        }
+    }
+    else
+    {
+        client->terminateWorkspaceDiagnostics();
+        client->refreshWorkspaceDiagnostics();
+    }
 }
 
 lsp::PartialResponse<lsp::WorkspaceDiagnosticReport> LanguageServer::workspaceDiagnostic(const lsp::WorkspaceDiagnosticParams& params)
