+++ src/operations/GotoDefinition.cpp
@@ -8,6 +8,8 @@
 
 lsp::DefinitionResult WorkspaceFolder::gotoDefinition(const lsp::DefinitionParams& params)
 {
+    ensureConfigured();
+
     lsp::DefinitionResult result{};
 
     auto moduleName = fileResolver.getModuleName(params.textDocument.uri);
@@ -87,7 +89,7 @@ lsp::DefinitionResult WorkspaceFolder::gotoDefinition(const lsp::DefinitionParam
         {
             if (definitionModuleName)
             {
-                if (auto file = fileResolver.resolveToRealPath(*definitionModuleName))
+                if (auto file = platform->resolveToRealPath(*definitionModuleName))
                 {
                     auto document = fileResolver.getTextDocumentFromModuleName(*definitionModuleName);
                     auto uri = document ? document->uri() : Uri::file(*file);
@@ -114,7 +116,7 @@ lsp::DefinitionResult WorkspaceFolder::gotoDefinition(const lsp::DefinitionParam
         {
             if (auto importedName = lookupImportedModule(*scope, reference->prefix.value().value))
             {
-                auto fileName = fileResolver.resolveToRealPath(*importedName);
+                auto fileName = platform->resolveToRealPath(*importedName);
                 if (!fileName)
                     return result;
                 uri = Uri::file(*fileName);
@@ -156,6 +158,8 @@ lsp::DefinitionResult WorkspaceFolder::gotoDefinition(const lsp::DefinitionParam
 
 std::optional<lsp::Location> WorkspaceFolder::gotoTypeDefinition(const lsp::TypeDefinitionParams& params)
 {
+    ensureConfigured();
+
     // If its a binding, we should find its assigned type if possible, and then find the definition of that type
     // If its a type, then just find the definintion of that type (i.e. the type alias)
 
@@ -194,7 +198,7 @@ std::optional<lsp::Location> WorkspaceFolder::gotoTypeDefinition(const lsp::Type
             {
                 if (auto importedName = lookupImportedModule(*scope, reference->prefix.value().value))
                 {
-                    auto fileName = fileResolver.resolveToRealPath(*importedName);
+                    auto fileName = platform->resolveToRealPath(*importedName);
                     if (!fileName)
                         return std::nullopt;
                     uri = Uri::file(*fileName);
