+++ src/operations/InlayHints.cpp
@@ -290,6 +290,8 @@ struct InlayHintVisitor : public Luau::AstVisitor
 
 lsp::InlayHintResult WorkspaceFolder::inlayHint(const lsp::InlayHintParams& params)
 {
+    ensureConfigured();
+
     auto config = client->getConfiguration(rootUri);
 
     auto moduleName = fileResolver.getModuleName(params.textDocument.uri);
