+++ src/operations/WorkspaceSymbol.cpp
@@ -105,6 +105,8 @@ struct WorkspaceSymbolsVisitor : public Luau::AstVisitor
 
 std::optional<std::vector<lsp::WorkspaceSymbol>> WorkspaceFolder::workspaceSymbol(const lsp::WorkspaceSymbolParams& params)
 {
+    ensureConfigured();
+
     std::vector<lsp::WorkspaceSymbol> result;
 
     for (const auto& [moduleName, sourceModule] : frontend.sourceModules)
@@ -120,7 +122,7 @@ std::optional<std::vector<lsp::WorkspaceSymbol>> WorkspaceFolder::workspaceSymbo
         }
         else
         {
-            if (auto filePath = fileResolver.resolveToRealPath(moduleName))
+            if (auto filePath = platform->resolveToRealPath(moduleName))
             {
                 if (auto source = fileResolver.readSource(moduleName))
                 {
