+++ src/platform/roblox/RobloxSourceNode.cpp
@@ -1,13 +1,9 @@
-#include <optional>
-#include <filesystem>
-#include "LSP/Sourcemap.hpp"
-#include "LSP/Utils.hpp"
+#include "Platform/RobloxPlatform.hpp"
 
 bool SourceNode::isScript()
 {
     return className == "ModuleScript" || className == "Script" || className == "LocalScript";
 }
-
 
 /// NOTE: Use `WorkspaceFileResolver::getRealPathFromSourceNode()` instead of this function where
 /// possible, as that will ensure it is relative to the correct workspace root.
@@ -65,61 +61,4 @@ std::optional<SourceNodePtr> SourceNode::findAncestor(const std::string& ancesto
         current = currentPtr->parent;
     }
     return std::nullopt;
-}
-
-Luau::SourceCode::Type sourceCodeTypeFromPath(const std::filesystem::path& requirePath)
-{
-    auto filename = requirePath.filename().generic_string();
-    if (endsWith(filename, ".server.lua") || endsWith(filename, ".server.luau"))
-    {
-        return Luau::SourceCode::Type::Script;
-    }
-    else if (endsWith(filename, ".client.lua") || endsWith(filename, ".client.luau"))
-    {
-        return Luau::SourceCode::Type::Local;
-    }
-
-    return Luau::SourceCode::Type::Module;
-}
-
-std::string jsonValueToLuau(const json& val)
-{
-    if (val.is_string() || val.is_number() || val.is_boolean())
-    {
-        return val.dump();
-    }
-    else if (val.is_null())
-    {
-        return "nil";
-    }
-    else if (val.is_array())
-    {
-        std::string out = "{";
-        for (auto& elem : val)
-        {
-            out += jsonValueToLuau(elem);
-            out += ";";
-        }
-
-        out += "}";
-        return out;
-    }
-    else if (val.is_object())
-    {
-        std::string out = "{";
-
-        for (auto& [key, value] : val.items())
-        {
-            out += "[\"" + key + "\"] = ";
-            out += jsonValueToLuau(value);
-            out += ";";
-        }
-
-        out += "}";
-        return out;
-    }
-    else
-    {
-        return ""; // TODO: should we error here?
-    }
 }
